Fix options for data sent to UI
This will allow screenshot_basic:requestScreenshot to actually work correctly. 
The original code would not receive the options, and would cause the server export "requestClientScreenshot" to not send any requests back to the server when fileName is nil. 
Since the readme states that an undefined fileName should return a data uri this is the incorrect behavior