Update main.ts
For this to work you will need an account on Imgur. Create an application and get the Client-ID which has to be put in ui/src/main.ts line 204.
THIS HAS TO BE DONE BEFORE RUNNING THE screenshot-basic for the first time!
Also change webhook url on every other resource that is using screenshot-basic so for example in qb-phone change the webhook option so it looks like this:
Config.Webhook = "imgur"